refactor : Split FlipCard from WordBookFragment and use dp unit instead of hardcoding
